tests/wpt/tests/css/css-sizing/aspect-ratio/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS aspect-ratio property</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio">
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<meta name="assert" content="Property does not inherit.">
<meta name="assert" content="Property has initial value according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_not_inherited('aspect-ratio', 'auto', '2 / 1');
</script>
</body>
</html>
